Fiddled around with node and edge creation. There are still static stores in DefaultNode, DefaultEdge, ShapeNode and ShapeEdge, but at least the store (and factory) functionality is delegated to dedicated classes NodeStore and EdgeStore.
I introduced HostGraph/HostNode types (I like this better than StateGraph/StateNode) but this is not yet connected to anything. We need to discuss (problems in) subtyping: I guess ShapeGraph/ShapeNode should be subtypes, but HostNode is an interface implemented by DefaultNode and ValueNode; how do we effectively subtype this?