Added full support for TermAlgebras.
This involved a cleanup of the Signature & Algebra functionality. Essentially, values can be represented in three different ways:

- The actual algebra representation
- The java representation (methods Algebra.toJavaValue and Algebra.toValueFromJava)
- Representation as a term (methods Algebra.toTerm and Algebra.toValue)

The representation as a string is gone, as term algebra values cannot be so represented.

Also added an Antlr parser for expressions, replacing the hand-crafted recursive descent parser in groove.grammar.aspect. Next step is to parse infix expressions.